pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markitdown-reference-image"
version = "0.1.0"
description = "Extract images from markdown files and highlight text chunks with bounding boxes"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Naveen Kumar Rajarajan", email = "smazeeapps@gmail.com"},
]
keywords = ["markitdown", "image", "reference"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "Pillow>=8.0.0",
    "markdown>=3.0.0",
    "beautifulsoup4>=4.9.0",
    "selenium>=4.0.0",
    "webdriver-manager>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]

[project.scripts]
markitdown-extract = "markitdown_reference_image.__main__:main"

[project.urls]
Homepage = "https://github.com/yourusername/markitdown-reference-image"
Repository = "https://github.com/yourusername/markitdown-reference-image"
Issues = "https://github.com/yourusername/markitdown-reference-image/issues"

[tool.black]
line-length = 88
target-version = ['py37']

[tool.pytest.ini_options]
testpaths = ["markitdown_reference_image/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
